Revert "[10.0.x] Drupal testing enhancements. (#3249)"
This reverts commit a98331ca997ce33e39c957f1ad081953f8654cc5.
