+++ src/Robo/Commands/Tests/AllCommand.php
@@ -20,7 +20,6 @@ class AllCommand extends BltTasks {
     $this->invokeCommands([
       'tests:behat:run',
       'tests:phpunit:run',
-      'tests:drupal:run',
       'tests:security:check:updates',
       'tests:security:check:composer',
       'tests:frontend:run',
