+++ src/Robo/Commands/Tests/PhpUnitCommand.php
@@ -2,21 +2,15 @@
 
 namespace Acquia\Blt\Robo\Commands\Tests;
 
+use Acquia\Blt\Robo\BltTasks;
 use Acquia\Blt\Robo\Exceptions\BltException;
+use Robo\Contract\VerbosityThresholdInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * Defines commands in the "tests" namespace.
  */
-class PhpUnitCommand extends DrupalTestCommand {
-
-  /**
-   * An array that contains configuration to override /
-   * customize phpunit commands.
-   *
-   * @var array
-   */
-  protected $phpunitConfig;
+class PhpUnitCommand extends BltTasks {
 
   /**
    * Directory in which test logs and reports are generated.
@@ -33,160 +27,92 @@ class PhpUnitCommand extends DrupalTestCommand {
   protected $reportFile;
 
   /**
+   * An array that contains configuration to override /
+   * customize phpunit commands.
+   *
+   * @var array
+   */
+  protected $phpunitConfig;
+
+  /**
    * This hook will fire for all commands in this command file.
    *
    * @hook init
    */
   public function initialize() {
-    parent::initialize();
-    $this->createReportsDir();
-  }
-
-  /**
-   * Setup and run PHPUnit tests.
-   *
-   * @command tests:phpunit:run
-   * @aliases tpr phpunit tests:phpunit
-   * @description Executes all PHPUnit tests.
-   *
-   * @throws \Exception
-   *   Throws an exception if any test fails.
-   */
-  public function runPhpUnitTests() {
-    $this->reportsDir = $this->getConfigValue('tests.reports.localDir') . '/phpunit';
+    $this->reportsDir = $this->getConfigValue('reports.localDir') . '/phpunit';
     $this->reportFile = $this->reportsDir . '/results.xml';
-    $this->phpunitConfig = $this->getConfigValue('tests.phpunit');
-    try {
-      $this->executeTests();
-    }
-    catch (\Exception $e) {
-      throw $e;
-    }
-  }
-
-  /**
-   * Setup and run Drupal tests.
-   *
-   * @command tests:drupal:phpunit:run
-   * @description Executes all PHPUnit tests. Launches chromedriver prior to execution.
-   * @hidden
-   *
-   * @throws \Exception
-   *   Throws an exception if any test fails.
-   */
-  public function runDrupalTests() {
-    $this->reportsDir = $this->getConfigValue('tests.reports.localDir') . '/drupal/phpunit';
-    $this->reportFile = $this->reportsDir . '/results.xml';
-    $this->phpunitConfig = $this->getConfigValue('tests.drupal.phpunit');
-    if ($this->drupalTestRunner == 'phpunit') {
-      try {
-        parent::run();
-      }
-      catch (\Exception $e) {
-        throw $e;
-      }
-    }
+    $this->testsDir = $this->getConfigValue('repo.root') . '/tests/phpunit';
+    $this->phpunitConfig = $this->getConfigValue('phpunit');
   }
 
   /**
    * Executes all PHPUnit tests.
+   *
+   * @command tests:phpunit:run
+   * @aliases tpr phpunit tests:phpunit
    */
-  public function executeTests() {
-    if (is_array($this->phpunitConfig)) {
-      foreach ($this->phpunitConfig as $test) {
-        $task = $this->taskPhpUnitTask()
-          ->xml($this->reportFile)
-          ->printOutput(TRUE)
-          ->printMetadata(FALSE);
+  public function testsPhpUnit() {
+    $this->createLogs();
+    foreach ($this->phpunitConfig as $test) {
+      $task = $this->taskPHPUnit()
+        ->xml($this->reportFile)
+        ->printOutput(TRUE)
+        ->printMetadata(FALSE);
 
-        if (isset($test['path'])) {
-          $task->dir($test['path']);
-        }
-
-        if ($this->output()->getVerbosity() >= OutputInterface::VERBOSITY_NORMAL) {
-          $task->printMetadata(TRUE);
-          $task->verbose();
-        }
-
-        if (isset($this->testingEnvString)) {
-          $task->testEnvVars($this->testingEnvString);
-        }
-
-        if (isset($this->apacheRunUser)) {
-          $task->user($this->apacheRunUser);
-        }
-
-        if (isset($this->sudoRunTests) && ($this->sudoRunTests)) {
-          $task->sudo();
-        }
-
-        if (isset($test['bootstrap'])) {
-          $task->bootstrap($test['bootstrap']);
-        }
-
-        if (isset($test['config'])) {
-          $task->configFile($test['config']);
-        }
-
-        if (isset($test['debug']) && ($test['debug'])) {
-          $task->debug();
-        }
-
-        if (isset($test['exclude'])) {
-          $task->excludeGroup($test['exclude']);
-        }
-
-        if (isset($test['filter'])) {
-          $task->filter($test['filter']);
-        }
-
-        if (isset($test['group'])) {
-          $task->group($test['group']);
-        }
-
-        if (isset($test['printer'])) {
-          $task->printer($test['printer']);
-        }
-
-        if (isset($test['stop-on-error']) && ($test['stop-on-error'])) {
-          $task->stopOnError();
-        }
-
-        if (isset($test['stop-on-failure']) && ($test['stop-on-failure'])) {
-          $task->stopOnFailure();
-        }
-
-        if (isset($test['testdox']) && ($test['testdox'])) {
-          $task->testdox();
-        }
-
-        if (isset($test['class'])) {
-          $task->arg($test['class']);
-          if (isset($test['file'])) {
-            $task->arg($test['file']);
-          }
-        }
-        else {
-          if (isset($test['directory'])) {
-            $task->arg($test['directory']);
-          }
-        }
-
-        if ((isset($test['testsuites']) && is_array($test['testsuites'])) || isset($test['testsuite'])) {
-          if (isset($test['testsuites'])) {
-            $task->testsuite(implode(',', $test['testsuites']));
-          }
-          elseif (isset($test['testsuite'])) {
-            $task->testsuite($test['testsuite']);
-          }
-        }
-
-        $result = $task->run();
-        if (!$result->wasSuccessful()) {
-          throw new BltException("PHPUnit tests failed.");
+      if (isset($test['class'])) {
+        $task->arg($test['class']);
+        if (isset($test['file'])) {
+          $task->arg($test['file']);
         }
       }
+      else {
+        if (isset($test['path'])) {
+          $task->arg($test['path']);
+        }
+      }
+
+      if (isset($test['path'])) {
+        $task->dir($test['path']);
+      }
+
+      if ($this->output()->getVerbosity() >= OutputInterface::VERBOSITY_NORMAL) {
+        $task->printMetadata(TRUE);
+        $task->arg('-v');
+      }
+
+      $supported_options = [
+        'config' => 'configuration',
+        'exclude-group' => 'exclude-group',
+        'filter' => 'filter',
+        'group' => 'group',
+        'testsuite' => 'testsuite',
+      ];
+
+      foreach ($supported_options as $yml_key => $option) {
+        if (isset($test[$yml_key])) {
+          $task->option("--$option", $test[$yml_key]);
+        }
+      }
+
+      $result = $task->run();
+      $exit_code = $result->getExitCode();
+
+      if ($exit_code) {
+        throw new BltException("PHPUnit tests failed.");
+      }
     }
+  }
+
+  /**
+   * Creates empty log directory and log file for PHPUnit tests.
+   */
+  protected function createLogs() {
+    $this->taskFilesystemStack()
+      ->mkdir($this->reportsDir)
+      ->touch($this->reportFile)
+      ->setVerbosityThreshold(VerbosityThresholdInterface::VERBOSITY_VERBOSE)
+      ->run();
   }
 
 }
