+++ template/blt/ci.blt.yml
@@ -6,7 +6,6 @@
 # Relative to the drupal docroot directory.
 # setup.dump-file: /tmp/my-dump-file.sql
 tests.run-server: true
-tests.drupal.sudo-run-tests: false
 # The local.hostname must be set to 127.0.0.1:8888 because we are using drush runserver to test the site.
 project.local.hostname: 127.0.0.1:8888
 drush.debug: false
