+++ tests/phpunit/Blt/TestPhpUnitCommandTest.php
@@ -10,36 +10,10 @@ use Acquia\Blt\Tests\BltProjectTestBase;
 class TestPhpUnitCommandTest extends BltProjectTestBase {
 
   /**
-   * @var string
-   */
-  protected $docroot;
-
-  /**
-   * @var string
-   */
-  protected $reporoot;
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setUp() {
-    parent::setUp();
-    $this->blt('recipes:blt:init:command');
-    $this->docroot = $this->config->get("docroot");
-    $this->reporoot = $this->config->get("repo.root");
-  }
-
-  /**
    * Test that ExampleTest.php is correctly executed and passes.
    */
   public function testPhpUnitCommandExampleTests() {
-    list($status_code, $output, $config) = $this->blt("tests:phpunit:run", [
-      "--define" => [
-        "tests.phpunit.0.config=$this->docroot/core/phpunit.xml.dist",
-        "tests.phpunit.0.path=$this->reporoot/tests/phpunit",
-        "tests.phpunit.0.directory=$this->reporoot/tests/phpunit",
-      ],
-    ]);
+    list($status_code, $output, $config) = $this->blt("tests:phpunit:run");
     $results = file_get_contents($this->sandboxInstance . '/reports/phpunit/results.xml');
     $this->assertContains('tests="1" assertions="1"', $results);
     $this->assertContains('name="testExample" class="My\Example\Project\Tests\ExampleTest"', $results);
@@ -50,13 +24,7 @@ class TestPhpUnitCommandTest extends BltProjectTestBase {
    */
   public function testPhpUnitCommandNoTests() {
     $this->fs->remove($this->sandboxInstance . "/tests/phpunit/ExampleTest.php");
-    list($status_code, $output, $config) = $this->blt("tests:phpunit:run", [
-      "--define" => [
-        "tests.phpunit.0.config=$this->docroot/core/phpunit.xml.dist",
-        "tests.phpunit.0.path=$this->reporoot/tests/phpunit",
-        "tests.phpunit.0.directory=$this->reporoot/tests/phpunit",
-      ],
-    ]);
+    list($status_code, $output, $config) = $this->blt("tests:phpunit:run");
     $this->assertEquals(0, $status_code);
   }
